Fix circle hitboxes

+++ f_input.pde
@@ -3,8 +3,8 @@ void mouseWheel(MouseEvent event) {
   	scroll += event.getCount();
 }
 
-int mx;
-int my;
+float mx;
+float my;
 // set mx and my to mouse position upon clicking
 void mouseClicked() {
   	mx = mouseX;
